+++ src/main/java/kitchenpos/menu/domain/MenuProducts.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.menu.domain;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -6,7 +6,7 @@ import java.util.stream.Collectors;
 import javax.persistence.CascadeType;
 import javax.persistence.Embeddable;
 import javax.persistence.OneToMany;
-import kitchenpos.dto.MenuProductResponse;
+import kitchenpos.menu.dto.MenuProductResponse;
 
 @Embeddable
 public class MenuProducts {
