+++ src/main/java/kitchenpos/menu/dto/MenuResponse.java
@@ -1,7 +1,7 @@
-package kitchenpos.dto;
+package kitchenpos.menu.dto;
 
 import java.util.List;
-import kitchenpos.domain.Menu;
+import kitchenpos.menu.domain.Menu;
 
 public class MenuResponse {
     private Long id;
