+++ src/main/java/kitchenpos/menu/repository/MenuRepository.java
@@ -1,6 +1,6 @@
-package kitchenpos.repository;
+package kitchenpos.menu.repository;
 
-import kitchenpos.domain.Menu;
+import kitchenpos.menu.domain.Menu;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MenuRepository extends JpaRepository<Menu, Long> {
