+++ src/test/java/kitchenpos/acceptance/OrderAcceptanceTest.java
@@ -10,10 +10,10 @@ import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 import kitchenpos.domain.OrderStatus;
-import kitchenpos.dto.MenuResponse;
 import kitchenpos.dto.OrderResponse;
 import kitchenpos.fixture.acceptance.AcceptanceTestMenuFixture;
 import kitchenpos.fixture.acceptance.AcceptanceTestOrderTableFixture;
+import kitchenpos.menu.dto.MenuResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
