+++ src/test/java/kitchenpos/fixture/acceptance/AcceptanceTestMenuFixture.java
@@ -3,8 +3,8 @@ package kitchenpos.fixture.acceptance;
 import java.math.BigDecimal;
 import java.util.HashMap;
 import java.util.Map;
-import kitchenpos.acceptance.MenuAcceptanceTest;
-import kitchenpos.dto.MenuResponse;
+import kitchenpos.menu.acceptance.MenuAcceptanceTest;
+import kitchenpos.menu.dto.MenuResponse;
 import kitchenpos.product.domain.Product;
 
 public class AcceptanceTestMenuFixture {
