+++ src/test/java/kitchenpos/menu/domain/MenuProductsTest.java
@@ -1,10 +1,12 @@
-package kitchenpos.domain;
+package kitchenpos.menu.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.Arrays;
 import java.util.stream.Collectors;
-import kitchenpos.dto.MenuProductResponse;
+import kitchenpos.domain.Price;
+import kitchenpos.domain.Quantity;
+import kitchenpos.menu.dto.MenuProductResponse;
 import org.junit.jupiter.api.Test;
 
 class MenuProductsTest {
